rename get charts controller
