+++ controllers/web/get-charts-controller.js
