+++ index.js
@@ -70,7 +70,7 @@ const itemsListController = require('./controllers/web/itemList')
 const itemsSaveController = require('./controllers/web/saveItems')
 const getUserStatsController = require('./controllers/web/getUserStats')
 const getUserStatsSubmitController = require('./controllers/web/getUserStatsSubmit')
-const getChartsController = require('./controllers/web/getChartsController')
+const getChartsController = require('./controllers/web/get-charts-controller')
 const getChartsOverallController = require('./controllers/web/getChartsOverallController')
 
 const loginController = require('./controllers/auth/loginController')
